src/Python/requirements.txt
flask>=2.0.1
flask_sqlalchemy
flask_login
sqlalchemy-iris
